Correções na pontuação

+++ Models/Score.py
@@ -1,6 +1,6 @@
 from Core.Database import Database
 from sqlalchemy.ext.declarative import declarative_base
-from sqlalchemy import Column, Integer, String
+from sqlalchemy import Column, Float, Integer, String
 
 
 Base = declarative_base()
@@ -11,7 +11,7 @@ class Score(Base):
 
     id = Column(Integer, primary_key=True)
     student_id = Column(String(50))
-    score = Column(Integer)
+    score = Column(Float)
 
     def __init__(self, student_id, score):
         self.student_id = student_id
